requirements.txt
astroid==2.1.0
autopep8==1.4.3
Django==2.1.5
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pbr==5.1.2
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.2.2
pytz==2018.9
six==1.12.0
stevedore==1.30.0
virtualenv==16.3.0
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.2
wrapt==1.11.0
